requirements.txt
agno
python-dotenv
yfinance
pandas
numpy
Groq
duckduckgo-search
ddgs
openai